refactor(kanban): broken down components
