+++ lib/constants/kanban.ts
@@ -1,0 +1,15 @@
+/**
+ * Constants data for kanban UI
+ */
+
+// Column Colors
+export const PREDEFINED_COLORS = [
+  "#3B82F6",
+  "#10B981",
+  "#F59E0B",
+  "#EF4444",
+  "#8B5CF6",
+  "#06B6D4",
+  "#84CC16",
+  "#F97316",
+];
